refactor: Simplify and refactor DeepSeek-V4
